requirements.txt
phidata
openai
psycopg[binary]
pgvector
python-dotenv
nest-asyncio
duckduckgo-search
